+++ ajax/dropdownAllItems.php
@@ -56,22 +56,27 @@ if ($_POST["idtable"] && class_exists($_POST["idtable"])) {
 
     $field_id = Html::cleanId("dropdown_" . $_POST["name"] . $rand);
 
-    $p        = [
+    $displaywith = ['otherserial', 'serial'];
+    $p = [
         'value'               => 0,
         'valuename'           => Dropdown::EMPTY_VALUE,
         'itemtype'            => $_POST["idtable"],
         'display_emptychoice' => true,
-        'displaywith'         => ['otherserial', 'serial'],
-        '_idor_token'         => Session::getNewIDORToken($_POST["idtable"]),
+        'displaywith'         => $displaywith,
+    ];
+    $idor_params = [
+        'displaywith' => $displaywith,
     ];
     if (isset($_POST['value'])) {
         $p['value'] = $_POST['value'];
     }
     if (isset($_POST['entity_restrict'])) {
-        $p['entity_restrict'] = $_POST['entity_restrict'];
+        $p['entity_restrict']           = $_POST['entity_restrict'];
+        $idor_params['entity_restrict'] = $_POST['entity_restrict'];
     }
     if (isset($_POST['condition'])) {
-        $p['condition'] = $_POST['condition'];
+        $p['condition']           = $_POST['condition'];
+        $idor_params['condition'] = $_POST['condition'];
     }
     if (isset($_POST['used'])) {
         $_POST['used'] = Toolbox::jsonDecode($_POST['used'], true);
@@ -82,6 +87,7 @@ if ($_POST["idtable"] && class_exists($_POST["idtable"])) {
     if (isset($_POST['width'])) {
         $p['width'] = $_POST['width'];
     }
+    $p['_idor_token'] = Session::getNewIDORToken($_POST["idtable"], $idor_params);
 
     echo  Html::jsAjaxDropdown(
         $_POST["name"],