+++ ajax/socket.php
@@ -39,7 +39,7 @@ include('../inc/includes.php');
 header("Content-Type: text/html; charset=UTF-8");
 Html::header_nocache();
 
-Session::checkLoginUser();
+Session::checkCentralAccess();
 
 switch ($_POST['action']) {
     case 'getItemsFromItemtype':